+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* Fix issue where customised schema validation message was incorrect for a modified payload
+
 # 5.0.1
 
 * Improve speed of random schema generation for customised content items